+++ lib/ts/querier.ts
@@ -21,6 +21,8 @@ import { RATE_LIMIT_STATUS_CODE } from "./constants";
 import { logDebugMessage } from "./logger";
 import { UserContext } from "./types";
 import { NetworkInterceptor } from "./types";
+import { QuerierError } from "./QuerierError";
+import SuperTokens from "./supertokens";
 
 export class Querier {
     private static initCalled = false;
@@ -48,11 +50,26 @@ export class Querier {
         this.rIdToCore = rIdToCore;
     }
 
-    getAPIVersion = async (): Promise<string> => {
+    getAPIVersion = async (userContext: UserContext): Promise<string> => {
         if (Querier.apiVersion !== undefined) {
             return Querier.apiVersion;
         }
         ProcessState.getInstance().addState(PROCESS_STATE.CALLING_SERVICE_IN_GET_API_VERSION);
+        const st = SuperTokens.getInstanceOrThrowError();
+        const appInfo = st.appInfo;
+
+        const queryParamsObj: {
+            apiDomain: string;
+            websiteDomain?: string;
+        } = {
+            apiDomain: appInfo.apiDomain.getAsStringDangerous(),
+        };
+        const request = st.getRequestFromUserContext(userContext);
+        if (request !== undefined) {
+            queryParamsObj.websiteDomain = appInfo.getOrigin({ request, userContext }).getAsStringDangerous();
+        }
+        const queryParams = new URLSearchParams(queryParamsObj).toString();
+
         let { body: response } = await this.sendRequestHelper(
             new NormalisedURLPath("/apiversion"),
             "GET",
@@ -63,7 +80,7 @@ export class Querier {
                         "api-key": Querier.apiKey,
                     };
                 }
-                let response = await doFetch(url, {
+                let response = await doFetch(url + `?${queryParams}`, {
                     method: "GET",
                     headers,
                 });
@@ -130,7 +147,7 @@ export class Querier {
             path,
             "POST",
             async (url: string) => {
-                let apiVersion = await this.getAPIVersion();
+                let apiVersion = await this.getAPIVersion(userContext);
                 let headers: any = {
                     "cdi-version": apiVersion,
                     "content-type": "application/json; charset=utf-8",
@@ -187,7 +204,7 @@ export class Querier {
             path,
             "DELETE",
             async (url: string) => {
-                let apiVersion = await this.getAPIVersion();
+                let apiVersion = await this.getAPIVersion(userContext);
                 let headers: any = { "cdi-version": apiVersion, "content-type": "application/json; charset=utf-8" };
                 if (Querier.apiKey !== undefined) {
                     headers = {
@@ -247,7 +264,7 @@ export class Querier {
             path,
             "GET",
             async (url: string) => {
-                let apiVersion = await this.getAPIVersion();
+                let apiVersion = await this.getAPIVersion(userContext);
                 let headers: any = { "cdi-version": apiVersion };
                 if (Querier.apiKey !== undefined) {
                     headers = {
@@ -350,7 +367,7 @@ export class Querier {
             path,
             "GET",
             async (url: string) => {
-                let apiVersion = await this.getAPIVersion();
+                let apiVersion = await this.getAPIVersion(userContext);
                 let headers: any = { "cdi-version": apiVersion };
                 if (Querier.apiKey !== undefined) {
                     headers = {
@@ -402,7 +419,7 @@ export class Querier {
             path,
             "PUT",
             async (url: string) => {
-                let apiVersion = await this.getAPIVersion();
+                let apiVersion = await this.getAPIVersion(userContext);
                 let headers: any = { "cdi-version": apiVersion, "content-type": "application/json; charset=utf-8" };
                 if (Querier.apiKey !== undefined) {
                     headers = {
@@ -542,16 +559,19 @@ export class Querier {
                     }
                 }
 
-                throw new Error(
+                const errorMessageFromCore = await err.text();
+
+                const errorText =
                     "SuperTokens core threw an error for a " +
-                        method +
-                        " request to path: '" +
-                        path.getAsStringDangerous() +
-                        "' with status code: " +
-                        err.status +
-                        " and message: " +
-                        (await err.text())
-                );
+                    method +
+                    " request to path: '" +
+                    path.getAsStringDangerous() +
+                    "' with status code: " +
+                    err.status +
+                    " and message: " +
+                    errorMessageFromCore;
+
+                throw new QuerierError(errorText, err.status as number, errorMessageFromCore);
             }
 
             throw err;