+++ lib/ts/recipe/dashboard/api/implementation.ts
@@ -48,7 +48,9 @@ export default function getAPIImplementation(): APIInterface {
             }
 
             let isSearchEnabled = false;
-            const cdiVersion = await Querier.getNewInstanceOrThrowError(input.options.recipeId).getAPIVersion();
+            const cdiVersion = await Querier.getNewInstanceOrThrowError(input.options.recipeId).getAPIVersion(
+                input.userContext
+            );
             if (maxVersion("2.20", cdiVersion) === cdiVersion) {
                 // Only enable search if CDI version is 2.20 or above
                 isSearchEnabled = true;