+++ CHANGELOG.md
@@ -29,6 +29,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+-   Add Multitenancy APIs dashboard APIs for:
+
+    -   Fetching a tenant's info
+    -   Deleting a tenant
+    -   Listing all tenants with their user counts
+    -   Creating or updating an existing tenant
+    -   Associating and disassociating a user from a tenant
+    -   Creating or updating third party config for a tenant
+    -   Deleting third party config for a tenant
+
 -   `passwordResetPOST`:
     -   now verifies the email address in all cases if the EmailVerification recipe is initialized
     -   now tries to link accounts based on account info if AccountLinking is enabled